requirements/run-requirements.txt
django==3.2.18
gunicorn==20.1.*
djangorestframework-simplejwt
djangorestframework-jwt
djangorestframework
djangorestframework-swagger
django-organizations
django-cache-memoize
django-postgres-extra
django_extensions
drf_yasg
psycopg2-binary
django-rest-swagger
Authlib==0.14.3
django-cors-headers
celery
redis
django-jsonfield
django-celery-beat
scipy
web3
drf-spectacular
websocket-client==1.5.1